Accept Enter key in keyInPause
    readlineSync.keyInPause('All done.');

generates a line that reads:

    All done. (Hit any key)

Before this commit, the Enter key was not accepted to continue from the
prompt. This could be a surprise for many users who naturally hit Enter
when they are prompted to hit any key. This commit makes it so
`keyInPause` accepts Enter, while being careful to not break other
`keyIn` methods.

`keyIn` accepts an option, `limit`, which makes it so that when
prompted, the program will not continue until the user has hit one of
the keys specified in `limit`. A previous commit, 6dffd8, made sure that
Enter was not accepted in any `keyIn` method so that the user could not
accidentally circumvent a `keyIn` with `limit` set.  This commit
introduces a new private option, `allowEmpty`. When and only when
`allowEmpty` is set to true, `keyIn` will accept Enter. `keyInPause` sets
this new option to `true`, allowing the user to hit Enter to move past a
`keyInPause` prompt.

Outside of `keyInPause`, `allowEmpty` defaults to false. This
`allowEmpty` option is not documented because it only makes sense in the
`keyInPause` scenario.